example_services/grpc_server/requirements.txt
grpcio-tools
grpcio
protobuf